Renamed VirtualSelect Provider Folder and NameSpace
