+++ src/scripts/OSUIFramework/Pattern/Dropdown/scss/_dropdown.scss
@@ -4,7 +4,7 @@
 /// Patterns - Interaction - Dropdown
 
 /// Import VirtualSelect provider library styles
-@import '../../../../Providers/Dropdown/Virtual_Select/scss/virtualselect';
+@import '../../../../Providers/Dropdown/VirtualSelect/scss/virtualselect';
 
 /// Dropdown Option Item image
 .osui-dropdown {
