+++ src/scripts/Providers/Dropdown/VirtualSelect/IVirtualSelect.ts
@@ -1,5 +1,5 @@
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-namespace Providers.Dropdown.Virtual_Select {
+namespace Providers.Dropdown.VirtualSelect {
 	/**
 	 * Defines the interface for VirtualSelect provider based Patterns
 	 */
