+++ src/scripts/Providers/Dropdown/VirtualSelect/Search/Enum.ts
@@ -1,5 +1,5 @@
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-namespace Providers.Dropdown.Virtual_Select.Search.Enum {
+namespace Providers.Dropdown.VirtualSelect.Search.Enum {
 	/**
 	 * Properties
 	 */
